Update os.cpp
I believe my logic in the drumint, tro and swapper functions is all correct. I'm sure there is a flaw or two in the code that im am not aware of, For Example: job is an iterator for the list of objects ( job table ). By dereferencing, the iterator, I believe that i can access the data members of job directly. ( Not 100% sure on this, but I cant run the code in its current state.)